# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sieve API",
    "summary": "Evidence-backed dependency vulnerability triage.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/orgs/{login}/overview": {
      "get": {
        "tags": [
          "organizations"
        ],
        "summary": "Overview",
        "operationId": "overview_api_v1_orgs__login__overview_get",
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orgs/{login}/repositories": {
      "get": {
        "tags": [
          "organizations"
        ],
        "summary": "Repositories",
        "operationId": "repositories_api_v1_orgs__login__repositories_get",
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RepositoryOut"
                  },
                  "title": "Response Repositories Api V1 Orgs  Login  Repositories Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orgs/{login}/findings": {
      "get": {
        "tags": [
          "organizations"
        ],
        "summary": "Findings",
        "operationId": "findings_api_v1_orgs__login__findings_get",
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Login"
            }
          },
          {
            "name": "verdict",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Verdict"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Verdict"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Severity"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "kev",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kev"
            }
          },
          {
            "name": "package",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Package"
            }
          },
          {
            "name": "repository_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Repository Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "presence",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Presence",
              "default": "present"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(risk|severity|discovered)$",
              "default": "risk",
              "title": "Sort"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orgs/{login}/activity": {
      "get": {
        "tags": [
          "organizations"
        ],
        "summary": "Activity",
        "operationId": "activity_api_v1_orgs__login__activity_get",
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Login"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Activity Api V1 Orgs  Login  Activity Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orgs/{login}/vex": {
      "get": {
        "tags": [
          "organizations"
        ],
        "summary": "Vex Documents",
        "operationId": "vex_documents_api_v1_orgs__login__vex_get",
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VexDocumentOut"
                  },
                  "title": "Response Vex Documents Api V1 Orgs  Login  Vex Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orgs/{login}/policy": {
      "get": {
        "tags": [
          "organizations"
        ],
        "summary": "Get Policy",
        "operationId": "get_policy_api_v1_orgs__login__policy_get",
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "organizations"
        ],
        "summary": "Put Policy",
        "operationId": "put_policy_api_v1_orgs__login__policy_put",
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Login"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orgs/{login}/search": {
      "get": {
        "tags": [
          "organizations"
        ],
        "summary": "Search",
        "operationId": "search_api_v1_orgs__login__search_get",
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Login"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditEventOut": {
        "properties": {
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "actor_type": {
            "type": "string",
            "title": "Actor Type"
          },
          "actor_login": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Login"
          },
          "target_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Type"
          },
          "target_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          }
        },
        "type": "object",
        "required": [
          "seq",
          "action",
          "actor_type",
          "actor_login",
          "target_type",
          "target_id",
          "data",
          "correlation_id",
          "occurred_at"
        ],
        "title": "AuditEventOut"
      },
...
      "FindingPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FindingItemOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "next_cursor"
        ],
        "title": "FindingPage"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrganizationOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "login": {
            "type": "string",
            "title": "Login"
          },
          "account_type": {
            "type": "string",
            "title": "Account Type"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "type": "object",
        "required": [
          "id",
          "login",
          "account_type",
          "role"
        ],
        "title": "OrganizationOut"
      },
      "OverviewOut": {
        "properties": {
          "organization": {
            "$ref": "#/components/schemas/OrganizationOut"
          },
          "totals": {
            "$ref": "#/components/schemas/VerdictCounts"
          },
          "severity": {
            "items": {
              "$ref": "#/components/schemas/SeverityBreakdown"
            },
            "type": "array",
            "title": "Severity"
          },
          "kev_findings": {
            "items": {
              "$ref": "#/components/schemas/FindingItemOut"
            },
            "type": "array",
            "title": "Kev Findings"
          },
          "top_findings": {
            "items": {
              "$ref": "#/components/schemas/FindingItemOut"
            },
            "type": "array",
            "title": "Top Findings"
          },
          "repositories": {
            "items": {
              "$ref": "#/components/schemas/RepositoryOut"
            },
            "type": "array",
            "title": "Repositories"
          },
          "activity": {
            "items": {
              "$ref": "#/components/schemas/AuditEventOut"
            },
            "type": "array",
            "title": "Activity"
          }
        },
        "type": "object",
        "required": [
          "organization",
          "totals",
          "severity",
          "kev_findings",
          "top_findings",
          "repositories",
          "activity"
        ],
        "title": "OverviewOut"
      },
...
      "PolicyIn": {
        "properties": {
          "rules": {
            "additionalProperties": true,
            "type": "object",
            "title": "Rules"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rules"
        ],
        "title": "PolicyIn"
      },
      "PolicyOut": {
        "properties": {
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "rules": {
            "additionalProperties": true,
            "type": "object",
            "title": "Rules"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "history": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "History"
          },
          "can_edit": {
            "type": "boolean",
            "title": "Can Edit"
          }
        },
        "type": "object",
        "required": [
          "version",
          "rules",
          "created_at",
          "history",
          "can_edit"
        ],
        "title": "PolicyOut"
      },
      "Presence": {
        "type": "string",
        "enum": [
          "present",
          "resolved"
        ],
        "title": "Presence"
      },
...
      "RepositoryOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "default_branch": {
            "type": "string",
            "title": "Default Branch"
          },
          "is_private": {
            "type": "boolean",
            "title": "Is Private"
          },
          "latest_scan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScanOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "counts": {
            "$ref": "#/components/schemas/VerdictCounts"
          },
          "health": {
            "type": "string",
            "enum": [
              "healthy",
              "attention",
              "failing",
              "pending"
            ],
            "title": "Health"
          }
        },
        "type": "object",
        "required": [
          "id",
          "full_name",
          "source",
          "default_branch",
          "is_private",
          "latest_scan",
          "counts",
          "health"
        ],
        "title": "RepositoryOut"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "owner",
          "admin",
          "member",
          "viewer"
        ],
        "title": "Role"
      },
...
      "SearchOut": {
        "properties": {
          "repositories": {
            "items": {
              "$ref": "#/components/schemas/RepositoryOut"
            },
            "type": "array",
            "title": "Repositories"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingItemOut"
            },
            "type": "array",
            "title": "Findings"
          },
          "packages": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Packages"
          }
        },
        "type": "object",
        "required": [
          "repositories",
          "findings",
          "packages"
        ],
        "title": "SearchOut"
      },
      "Severity": {
        "type": "string",
        "enum": [
          "critical",
          "high",
          "medium",
          "low",
          "unknown"
        ],
        "title": "Severity"
      },
      "SeverityBreakdown": {
        "properties": {
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "reachable": {
            "type": "integer",
            "title": "Reachable"
          },
          "needs_review": {
            "type": "integer",
            "title": "Needs Review"
          },
          "not_reached": {
            "type": "integer",
            "title": "Not Reached"
          }
        },
        "type": "object",
        "required": [
          "severity",
          "reachable",
          "needs_review",
          "not_reached"
        ],
        "title": "SeverityBreakdown"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Verdict": {
        "type": "string",
        "enum": [
          "reachable",
          "not_reached",
          "needs_review"
        ],
        "title": "Verdict"
      },
      "VerdictCounts": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "reachable": {
            "type": "integer",
            "title": "Reachable",
            "default": 0
          },
          "needs_review": {
            "type": "integer",
            "title": "Needs Review",
            "default": 0
          },
          "not_reached": {
            "type": "integer",
            "title": "Not Reached",
            "default": 0
          },
          "pending": {
            "type": "integer",
            "title": "Pending",
            "default": 0
          },
          "kev": {
            "type": "integer",
            "title": "Kev",
            "default": 0
          }
        },
        "type": "object",
        "title": "VerdictCounts"
      },
      "VexDocumentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "format": {
            "$ref": "#/components/schemas/VexFormat"
          },
          "repository_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repository Id"
          },
          "repository_full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repository Full Name"
          },
          "statement_count": {
            "type": "integer",
            "title": "Statement Count"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "format",
          "repository_id",
          "repository_full_name",
          "statement_count",
          "sha256",
          "created_at"
        ],
        "title": "VexDocumentOut"
      },
      "VexFormat": {
        "type": "string",
        "enum": [
          "openvex",
          "cyclonedx"
        ],
        "title": "VexFormat"
      },
...
    }
  }
}